% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viability_IC50.R
\name{viability_IC50}
\alias{viability_IC50}
\title{Calculate IC50 of conditions tested in the viability assay}
\usage{
viability_IC50(template, viability_data)
}
\arguments{
\item{template}{Path to template file}

\item{viability_data}{Path to viability plate results .xlsx file}
}
\description{
Calculate IC50 of conditions tested in the viability assay
}
