% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viability_template.R
\name{viability_template}
\alias{viability_template}
\title{Generate a template of the plate used for a viability assay}
\usage{
viability_template(drug_max, conditions, save_to)
}
\arguments{
\item{drug_max}{Maximum drug concentration used for the assay}

\item{conditions}{\verb{<vector>} Conditions used in the assay.
Usually different cell lines or cells with different treatments.
4 conditions are expected. If less than 4 conditions were used, complete with \verb{<NA>}}

\item{save_to}{\verb{<string>} Path to file in which to save the template}
}
\description{
Generate a template of the plate used for a viability assay
}
